backend/requirements.txt
flask
flask-bcrypt
flask-jwt-extended
flask-cors
psycopg2-binary
python-dotenv
